requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
kafka-python==2.0.2
